Renamed the param-sweep script
